plated/requirements.txt
django
django-crispy-forms
crispy-bootstrap4
pillow
django-nested-admin
matplotlib
django-environ
